+++ core/execution/liquidation/engine_test.go
@@ -47,7 +47,6 @@ type tstEngine struct {
 	broker *bmocks.MockBroker
 	tSvc   *cmocks.MockTimeService
 	pos    *mocks.MockPositions
-	set    *mocks.MockSettlement
 }
 
 type marginStub struct {
@@ -95,7 +94,6 @@ func TestNetworkReducesOverTime(t *testing.T) {
 	eng.broker.EXPECT().SendBatch(SliceLenMatcher[events.Event](1 * len(closed))).Times(1)
 	eng.pos.EXPECT().RegisterOrder(gomock.Any(), gomock.Any()).Times(2 * len(closed))
 	eng.pos.EXPECT().Update(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(len(closed))
-	eng.set.EXPECT().AddTrade(gomock.Any()).Times(len(closed))
 	pos, parties, trades := eng.ClearDistressedParties(ctx, eng.idgen, closed, num.UintZero(), num.UintZero())
 	require.Equal(t, len(closed), len(trades))
 	require.Equal(t, len(closed), len(pos))
@@ -177,7 +175,6 @@ func TestNetworkReducesOverTime(t *testing.T) {
 		eng.broker.EXPECT().SendBatch(SliceLenMatcher[events.Event](1 * len(closed))).Times(1)
 		eng.pos.EXPECT().RegisterOrder(gomock.Any(), gomock.Any()).Times(len(closed) * 2)
 		eng.pos.EXPECT().Update(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(len(closed))
-		eng.set.EXPECT().AddTrade(gomock.Any()).Times(len(closed))
 		pos, parties, trades := eng.ClearDistressedParties(ctx, eng.idgen, closed, num.UintZero(), num.UintZero())
 		require.Equal(t, len(closed), len(trades))
 		require.Equal(t, len(closed), len(pos))
@@ -263,7 +260,6 @@ func testOrderbookHasNoVolume(t *testing.T) {
 	eng.broker.EXPECT().SendBatch(SliceLenMatcher[events.Event](1 * len(closed))).Times(1)
 	eng.pos.EXPECT().RegisterOrder(gomock.Any(), gomock.Any()).Times(len(closed) * 2)
 	eng.pos.EXPECT().Update(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(len(closed))
-	eng.set.EXPECT().AddTrade(gomock.Any()).Times(len(closed))
 	pos, parties, trades := eng.ClearDistressedParties(ctx, eng.idgen, closed, num.UintZero(), num.UintZero())
 	require.Equal(t, len(closed), len(trades))
 	require.Equal(t, len(closed), len(pos))
@@ -307,7 +303,6 @@ func testOrderbookFractionRounding(t *testing.T) {
 	eng.broker.EXPECT().SendBatch(SliceLenMatcher[events.Event](1 * len(closed))).Times(1)
 	eng.pos.EXPECT().RegisterOrder(gomock.Any(), gomock.Any()).Times(len(closed) * 2)
 	eng.pos.EXPECT().Update(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(len(closed))
-	eng.set.EXPECT().AddTrade(gomock.Any()).Times(len(closed))
 	pos, parties, trades := eng.ClearDistressedParties(ctx, eng.idgen, closed, num.UintZero(), num.UintZero())
 	require.Equal(t, len(closed), len(trades))
 	require.Equal(t, len(closed), len(pos))
@@ -352,7 +347,6 @@ func testOrderbookExceedsVolume(t *testing.T) {
 	eng.broker.EXPECT().SendBatch(SliceLenMatcher[events.Event](1 * len(closed))).Times(1)
 	eng.pos.EXPECT().RegisterOrder(gomock.Any(), gomock.Any()).Times(len(closed) * 2)
 	eng.pos.EXPECT().Update(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(len(closed))
-	eng.set.EXPECT().AddTrade(gomock.Any()).Times(len(closed))
 	pos, parties, trades := eng.ClearDistressedParties(ctx, eng.idgen, closed, num.UintZero(), num.UintZero())
 	require.Equal(t, len(closed), len(trades))
 	require.Equal(t, len(closed), len(pos))
@@ -393,7 +387,6 @@ func TestLegacySupport(t *testing.T) {
 	eng.broker.EXPECT().SendBatch(SliceLenMatcher[events.Event](1 * len(closed))).Times(1)
 	eng.pos.EXPECT().RegisterOrder(gomock.Any(), gomock.Any()).Times(len(closed) * 2)
 	eng.pos.EXPECT().Update(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(len(closed))
-	eng.set.EXPECT().AddTrade(gomock.Any()).Times(len(closed))
 	pos, parties, trades := eng.ClearDistressedParties(ctx, eng.idgen, closed, num.UintZero(), num.UintZero())
 	require.Equal(t, len(closed), len(trades))
 	require.Equal(t, len(closed), len(pos))
@@ -415,7 +408,6 @@ func TestLegacySupport(t *testing.T) {
 	eng.broker.EXPECT().SendBatch(SliceLenMatcher[events.Event](1 * len(closed))).Times(1)
 	eng.pos.EXPECT().RegisterOrder(gomock.Any(), gomock.Any()).Times(len(closed) * 2)
 	eng.pos.EXPECT().Update(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(len(closed))
-	eng.set.EXPECT().AddTrade(gomock.Any()).Times(len(closed))
 	pos, parties, trades = eng.ClearDistressedParties(ctx, eng.idgen, closed, num.UintZero(), num.UintZero())
 	require.Equal(t, len(closed), len(trades))
 	require.Equal(t, len(closed), len(pos))
@@ -447,7 +439,6 @@ func TestLegacySupport(t *testing.T) {
 	eng.broker.EXPECT().SendBatch(SliceLenMatcher[events.Event](1 * len(closed))).Times(1)
 	eng.pos.EXPECT().RegisterOrder(gomock.Any(), gomock.Any()).Times(len(closed) * 2)
 	eng.pos.EXPECT().Update(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(len(closed))
-	eng.set.EXPECT().AddTrade(gomock.Any()).Times(len(closed))
 	pos, parties, trades = eng.ClearDistressedParties(ctx, eng.idgen, closed, num.UintZero(), num.UintZero())
 	require.Equal(t, len(closed), len(trades))
 	require.Equal(t, len(closed), len(pos))
@@ -553,8 +544,7 @@ func getTestEngine(t *testing.T, marketID string, config *types.LiquidationStrat
 	broker := bmocks.NewMockBroker(ctrl)
 	tSvc := cmocks.NewMockTimeService(ctrl)
 	pe := mocks.NewMockPositions(ctrl)
-	se := mocks.NewMockSettlement(ctrl)
-	engine := liquidation.New(logging.NewDevLogger(), config, marketID, broker, book, as, tSvc, ml, pe, se)
+	engine := liquidation.New(logging.NewDevLogger(), config, marketID, broker, book, as, tSvc, ml, pe)
 	return &tstEngine{
 		Engine: engine,
 		ctrl:   ctrl,
@@ -565,7 +555,6 @@ func getTestEngine(t *testing.T, marketID string, config *types.LiquidationStrat
 		broker: broker,
 		tSvc:   tSvc,
 		pos:    pe,
-		set:    se,
 	}
 }